requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
django-filter==23.3
djangorestframework-simplejwt==5.3.0
dj-database-url==2.1.0

# Development tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3
ipython==8.17.2

# Database (PostgreSQL)
# psycopg2-binary==2.9.7  # conda로 설치
